projects/Fix-it/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
jira==3.5.1
requests==2.31.0
python-jose==3.3.0
flask-cors==4.0.0
pytest==7.4.3
coverage==7.3.2
flask-jwt-extended==4.6.0
bcrypt==4.1.2
flask-login==0.6.3
